refactor: rename balius-proto -> balius-core, nest schema under proto::v0
The crate's job has been "the Balius-owned wire schema", but the name
"balius-proto" advertised protobuf as the only thing it carries.
`balius-core` is a roomier name for what is really the foundational
types crate for the runtime <-> WASM interface — protobuf today, more
things possibly later.

The schema module also gets nested: `balius_proto::cardano` becomes
`balius_core::proto::v0::cardano`. The `v0` segment makes the versioning
explicit: this schema is wire-compatible with utxorpc-spec 0.17 by
construction (the entire point of this PR), so it is frozen for
pre-BigInt worker compat. A future breaking change to the schema goes
under `proto::v1`, leaving `v0` consumers untouched.

Mechanical changes:
 - Crate dir balius-proto/ -> balius-core/
 - Package name -> balius-core; description updated
 - src/cardano.rs -> src/proto/v0/cardano.rs (plus new mod.rs files)
 - Workspace member + all 3 dependent Cargo.tomls updated
   (balius-runtime, balius-sdk, examples/asteria-tracker)
 - Import paths: balius_proto::cardano -> balius_core::proto::v0::cardano
   across balius-sdk, balius-runtime, examples/asteria-tracker
 - Doc/comment references updated

No behavior change. cargo check --workspace --tests clean; 2/2 wire-compat
tests in balius-core; 9/9 lib tests in balius-runtime (incl. all 8 u5c
convert tests); e2e::faucet_claim still passes; only the pre-existing
u5c-chainsync::wallet_balance stale-wasm failure remains.
