+++ balius-core/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "balius-proto"
+name = "balius-core"
 version = "0.5.2"
 edition = "2021"
 
-description = "Balius-owned protobuf schema for the runtime <-> WASM interface"
+description = "Balius-owned core types and protocol schemas for the runtime <-> WASM interface"
 license = "Apache-2.0"
 repository = "https://github.com/txpipe/balius"
