+++ balius-core/src/lib.rs
@@ -1,0 +1,3 @@
+// Balius-owned types used across the runtime <-> WASM boundary.
+
+pub mod proto;
