+++ balius-core/src/proto/mod.rs
@@ -1,0 +1,5 @@
+// Versioned schemas for what crosses the WIT boundary. Each version is
+// a frozen snapshot — bump to a new submodule when introducing a
+// breaking change rather than mutating an existing one.
+
+pub mod v0;
