+++ balius-core/src/proto/v0/mod.rs
@@ -1,0 +1,5 @@
+// v0 schema — wire-compatible with utxorpc-spec 0.17.0
+// `utxorpc.v1alpha.cardano`. Workers built against pre-BigInt SDKs
+// decode these bytes unchanged.
+
+pub mod cardano;
