--- balius-proto/src/lib.rs
@@ -1,5 +1,0 @@
-// Balius-owned schema. Wire-compatible with utxorpc-spec 0.17.0
-// `utxorpc.v1alpha.cardano`. Workers built against pre-BigInt SDKs decode
-// these bytes unchanged.
-
-pub mod cardano;
