+++ balius-runtime/Cargo.toml
@@ -25,7 +25,7 @@ itertools = "0.14.0"
 async-trait = "0.1.83"
 utxorpc = { version = "0.13.0" }
 # utxorpc = { path = "../../../utxorpc/rust-sdk" }
-balius-proto = { version = "0.5.2", path = "../balius-proto" }
+balius-core = { version = "0.5.2", path = "../balius-core" }
 tokio-util = "0.7.12"
 prost = "0.13"
 object_store = { version = "0.12.0", features = ["fs"] }
