+++ examples/asteria-tracker/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 balius-sdk = { version = "0.5.2", path = "../../balius-sdk" }
 serde = { version = "1.0.204", features = ["derive"] }
 serde_json = "1.0"
-balius-proto = { version = "0.5.2", path = "../../balius-proto" }
+balius-core = { version = "0.5.2", path = "../../balius-core" }
 hex = "0.4.3"
 url = "2.5"
